requirements.txt
aiohttp==3.8.1
aiomysql==0.0.22
aiosignal==1.2.0
aiosqlite==0.17.0
altair==4.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bitarray==1.2.2
bitcoin-etl==1.5.2
black==22.1.0
bleach==4.1.0
blinker==1.4
# Editable Git install with no remote (bq-helper==0.4.1)
# -e /drive_data/test_proj/venv/src/bq-helper
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
confluent-kafka==1.8.2
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.26
cytoolz==0.11.2
dask==2021.12.0
databases==0.4.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
diagrams==0.21.1
Django==3.2.12
docstring-parser==0.12
entrypoints==0.4
eth-abi==3.0.0
eth-account==0.6.0
eth-hash==0.3.2
eth-keyfile==0.6.0
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.6.0b5
eth-typing==3.0.0
eth-utils==2.0.0
executing==0.8.3
Faker==13.3.4
fastapi==0.75.0
fastapi-pagination==0.9.1
filelock==3.4.2
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2021.11.1
geographiclib==1.52
geopy==2.2.0
gino==1.0.1
gino-starlette==0.1.4
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.2
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
graphviz==0.19.2
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
guppy3==3.1.2
h11==0.13.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
ipfshttpclient==0.8.0a2
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso8601==0.1.16
jedi==0.18.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kafka-python==2.0.2
kiwisolver==1.3.2
llvmlite==0.38.0
locket==0.2.1
lru-dict==1.1.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
motor==2.5.1
msgpack==1.0.3
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
ndg-httpsclient==0.5.1
nest-asyncio==1.5.4
netaddr==0.8.0
notebook==6.4.10
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
orjson==3.6.6
orm==0.1.9
ormar==0.10.25
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
partd==1.2.0
pathspec==0.9.0
pexpect==4.8.0
piccolo==0.34.0
pickleshare==0.7.5
Pillow==9.0.1
pkg_resources==0.0.0
platformdirs==2.5.1
plotly==5.6.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
proto-plus==1.20.0
protobuf==3.19.3
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.12.0
pydantic==1.9.0
pydeck==0.7.1
Pygments==2.11.2
pymongo==3.12.3
Pympler==1.0.1
PyMySQL==0.9.3
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyperf==2.3.1
pyperformance==1.0.5
pypika-tortoise==0.1.3
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
PythonTurtle==0.3.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
ray==1.10.0
redis==4.1.3
requests==2.20.0
requests-oauthlib==1.3.1
rlp==3.0.0
rsa==4.8
scipy==1.8.0
semantic-version==2.8.5
semver==2.13.0
Send2Trash==1.8.0
sh==1.14.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.3.24
sqlalchemy-views==0.3.1
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.17.1
streamlit==1.7.0
targ==0.3.6
tenacity==8.0.1
terminado==0.13.3
testpath==0.6.0
tk==0.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tortoise-orm==0.17.8
tqdm==4.62.3
traitlets==5.1.1
typesystem==0.4.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.24.3
uvicorn==0.17.6
validators==0.18.2
varint==1.0.2
watchdog==2.1.6
wcwidth==0.2.5
web3==5.26.0
webencodings==0.5.1
websockets==9.1
widgetsnbextension==3.6.0
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
